Implement mail, file upload, user and token functionality
The update enhances the functionality of the MailService by adding the PHPMailer package and enabling the sending of email. It implements file upload feature to a local directory and AWS S3. It also updates the User entity and UserController for user registration, activation, and login capabilities. Various helper and validator classes are introduced to assist with these functionalities.