Remove Timestamp entity and refactor middlewares
This commit deletes the Timestamp entity and modifies how middlewares are defined. It also includes changes in EntityManagerFactory to enhance the method for creating EntityManager. The CORS middleware has been refactored to provide response headers correctly and the middleware definition has been optimized in the application setup. Certain updates in the composer.lock file have been made as part of the routine package updates.